ugo/backend/requirement.txt
django
Pillow
django-versatileimagefield